requirements.txt
flask>=3.0.0
python-dotenv>=1.0.0
google-genai>=0.7.0
requests>=2.31.0
